size of int value
size of int value

+++ WebAPI/Services/registration.cs
@@ -39,7 +39,7 @@ namespace WebAPI.Services
                 Email = model.Email,
                 Type = model.Type,
                 jobType = model.jobType,
-                Mobile = Convert.ToInt32(model.PhoneNumber),
+                Mobile =Convert.ToInt64(model.PhoneNumber),
                 FacilityId = model.FacilityId,
                 Country = model.NationalityId.ToString()
